requirements.txt
aiogram>=3.0
python-dotenv>=0.19
requests>=2.26